Applied the Observer pattern to Cell and Unit.
Cells extend the GameSubject class and Units implement the GameEventObserver interface.
Units register and unregister GameSubjects to observe. GameSubject notify observers of Events via the Notify method. 

Currently, only the "MoveEvent" is really implemented at all. These events are generated in EntityLocController in the updateUnitLocation and addEntity methods. EntityLocController also tells Unit which Cells it should observe with the private method updateCellsUnitIsObserving. When a Unit sees a "MoveEvent" it simply writes to console that it has observed an event.

The Event class has 4 attributes: 
orginCell - The Cell where the Event was generated.
sourceEntity - The Entity which caused the event to happen.
targetEntity - The Entity being effected by the event. 
type - What type of event it is. (Currently only MoveEvents and AttackEvents.)