Added AttackAction.cs and GameEventLogic.cs. Renamed GameEvent/Event.cs to GameEvent/GameEvent.cs. Made changes to ActionController and ActionCommand to allow for AttackAction. Many misc. changes.
AttackAction.cs defines an AttackAction where a Unit will move to the target until the target is within the Unit's attack range. Then it will attack it.

GameEventLogic is where GameEvents will be processed. Originally, I wanted each GameEventObserver to be able to define how it reacted to events. They would usually react by inserting actions to their actionQueue. However, this would require ZRTSLogic to be referenced by ZRTSModel which would create a cyclical dependency. As a workaround, each GameEventObserver keeps a list of GameEvents it has seen and Controller calls GameEventLogic.proccessEvents(GameEventObserver, GameWorld) to process the GameEvents the GameEventObserver has seen.